// src/app/models/product.ts
export default class Product {
    idProduct: number;
    name: String;
    category: String;
    opis: String;
    price: number;
    brKom: number;
    url: String;
    idProizvoda: number;
    kolicina: number;
    isAdded: boolean;
    idKorpe : number;
    brand: String;
    deleted: boolean;

}